requirements.txt
fastapi
uvicorn
python-dotenv
motor
httpx
pydantic
websockets
